(PC-10260) Use webapp v2 url for sharing offers and in recaptcha webview
