+++ src/features/bookings/components/BookingPropertiesSection.native.test.tsx
@@ -68,6 +68,7 @@ async function renderBookingProperties(booking: Booking) {
           displayDmsRedirection: true,
           idCheckAddressAutocompletion: false,
           useAppSearch: true,
+          isWebappV2Enabled: false,
         }}
       />
     )
