+++ src/features/deeplinks/__mocks__/listener.ts
@@ -1,0 +1,1 @@
+export const useListenDeepLinksEffect = jest.fn()
