+++ src/features/venue/pages/__tests__/Venue.native.test.tsx
@@ -2,6 +2,7 @@ import React from 'react'
 
 import { useRoute } from '__mocks__/@react-navigation/native'
 import { venueResponseSnap } from 'features/venue/fixtures/venueResponseSnap'
+import { reactQueryProviderHOC } from 'tests/reactQueryProviderHOC'
 import { render, waitFor } from 'tests/utils'
 
 import { Venue } from '../Venue'
@@ -25,7 +26,7 @@ describe('<Venue />', () => {
 
 async function renderVenue(id: number) {
   useRoute.mockImplementation(() => ({ params: { id } }))
-  const wrapper = render(<Venue />)
+  const wrapper = render(<Venue />, { wrapper: ({ children }) => reactQueryProviderHOC(children) })
   await waitFor(() => wrapper.getByTestId('Page de détail du lieu'))
   return wrapper
 }
