+++ src/libs/environment/__mocks__/env.ts
@@ -1,0 +1,1 @@
+export { env } from './envFixtures'
