+++ src/libs/environment/__mocks__/envFixtures.ts
@@ -1,0 +1,63 @@
+import { Environment } from 'libs/environment/types'
+
+export const env: Environment = {
+  ACCESSIBILITY_LINK: 'https://passculture.accessibility',
+  ALGOLIA_APPLICATION_ID: 'algoliaAppId',
+  ALGOLIA_INDEX_NAME: 'algoliaIndexName',
+  ALGOLIA_SEARCH_API_KEY: 'algoliaApiKey',
+  ANDROID_APP_ID: 'app.android',
+  API_BASE_URL: 'http://localhost',
+  APP_SEARCH_ENDPOINT: 'appSearchEndpoint',
+  APP_SEARCH_KEY: 'appSearchKey',
+  APPS_FLYER_DEV_KEY: 'appsFlyerDevKey',
+  BATCH_API_KEY_WEB: 'BATCH_API_KEY_WEB',
+  BATCH_AUTH_KEY: 'BATCH_AUTH_KEY',
+  BATCH_SUBDOMAIN: 'BATCH_SUBDOMAIN',
+  BATCH_VAPID_PUBLIC_KEY: 'BATCH_VAPID_PUBLIC_KEY',
+  CGU_LINK: 'https://passculture.cgu',
+  CONTENTFUL_ACCESS_TOKEN: 'accessToken',
+  CONTENTFUL_ENVIRONMENT: 'environment',
+  CONTENTFUL_SPACE_ID: 'contentfulSpaceId',
+  COOKIES_POLICY_LINK: 'https://passculture.cookies',
+  CULTURAL_SURVEY_TYPEFORM_URL: 'BATCH_VAPID_PUBLIC_KEY',
+  DATA_PRIVACY_CHART_LINK: 'https://passculture.data-privacy-chart',
+  DOC_CGU_URL:
+    'https//docs.passculture.app/textes-normatifs/mentions-legales-et-conditions-generales-dutilisation-de-lapplication-pass-culture',
+  DOC_PERSONAL_DATA_URL:
+    'https://docs.passculture.app/textes-normatifs/charte-des-donnees-personnelles',
+  DSM_URL: '//www.demarches-simplifiees.fr/commencer/inscription-pass-culture',
+  ENABLE_WHY_DID_YOU_RENDER: false,
+  ENV: 'testing',
+  FAQ_LINK: 'https://passculture.faq',
+  FEATURE_FLIPPING_ONLY_VISIBLE_ON_TESTING: true,
+  FIREBASE_APIKEY: 'FIREBASE_APIKEY',
+  FIREBASE_APPID: 'BATCH_VAPID_PUBLIC_KEY',
+  FIREBASE_AUTHDOMAIN: 'FIREBASE_APPID',
+  FIREBASE_DYNAMIC_LINK: 'passcultureapptesting.page.link',
+  FIREBASE_MESSAGINGSENDERID: 'FIREBASE_MESSAGINGSENDERID',
+  FIREBASE_PROJECTID: 'FIREBASE_PROJECTID',
+  FIREBASE_STORAGEBUCKET: 'FIREBASE_STORAGEBUCKET',
+  ID_CHECK_API_URL: 'https://passculture.idcheck',
+  ID_CHECK_URL: 'https://id-check-unit-tests',
+  IOS_APP_ID: 'app.ios',
+  IOS_APP_STORE_ID: '1557887412',
+  PRIVACY_POLICY_LINK: 'https://passculture.privacy',
+  PUBLIC_URL: 'https://web.example.com',
+  RECOMMENDATION_ENDPOINT: 'https://recommmendation-endpoint',
+  RECOMMENDATION_TOKEN: 'recommmendation-token',
+  SENTRY_DSN: 'SENTRY_DSN',
+  SIGNIN_IDENTIFIER: '',
+  SIGNIN_PASSWORD: '',
+  SIGNUP_DATE: '',
+  SIGNUP_POSTAL_CODE: '',
+  SIGNUP_RANDOM_EMAIL: false,
+  SIGNUP_RANDOM_PASSWORD: false,
+  SITE_KEY: 'SITE_KEY',
+  SUPPORT_EMAIL_ADDRESS: 'support@test.passculture.app',
+  TMX_FPSERVER: 'test.pc',
+  TMX_ORGID: 'xyzt',
+  UNIVERSAL_LINK: 'app.passculture-testing.beta.gouv.fr',
+  URL_PREFIX: 'passculture',
+  WEBAPP_URL: 'https://webapp.example.com',
+  WEBAPP_V2_DOMAIN: 'webapp-v2.example.com',
+}
