+++ src/libs/environment/env.ts
@@ -2,6 +2,6 @@ import Config from '@bam.tech/react-native-config'
 
 import { parseBooleanVariables } from 'libs/environment/parseBooleanVariables'
 
-import { Environment } from '.'
+import { Environment } from './types'
 
 export const env = parseBooleanVariables(Config) as Environment
