+++ src/libs/environment/env.web.ts
@@ -2,7 +2,7 @@ import { omit } from 'lodash'
 
 import { parseBooleanVariables } from 'libs/environment/parseBooleanVariables'
 
-import { Environment } from '.'
+import { Environment } from './types'
 
 // TODO: FIXME: configure cors on backend and disable proxy in webpackDevServer.config.js#L110 when doing https://passculture.atlassian.net/browse/PC-9847
 export const env = (parseBooleanVariables(
