+++ src/libs/environment/parseBooleanVariables.test.ts
@@ -4,7 +4,6 @@ describe('parseBooleanVariables', () => {
   const mockedConfig = {
     ENV: 'staging',
     API_BASE_URL: 'your-api.com',
-    WEBSOCKET_ENDPOINT: 'websocket-endpoint',
     FEATURE_FLIPPING_ONLY_VISIBLE_ON_TESTING: 'false',
     ENABLE_WHY_DID_YOU_RENDER: 'true',
     SENTRY_DSN: 'sentry-dsn',
@@ -20,7 +19,6 @@ describe('parseBooleanVariables', () => {
   })
   it('should not touch strings other than "true" and "false"', () => {
     expect(convertedConfig.API_BASE_URL).toEqual('your-api.com')
-    expect(convertedConfig.WEBSOCKET_ENDPOINT).toEqual('websocket-endpoint')
     expect(convertedConfig.SENTRY_DSN).toEqual('sentry-dsn')
     expect(convertedConfig.URL_PREFIX).toEqual('passculture')
   })
